backend/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
frozenlist==1.4.1
geoip2==4.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
maxminddb==2.6.2
multidict==6.0.5
openai==1.40.3
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==71.0.4
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
ua-parser==0.18.0
urllib3==2.2.2
yarl==1.9.4
